feat (DPLAN-11982): add new class ImageManager. Extract the handling of images in segment export in new class.
