+++ demosplan/DemosPlanCoreBundle/Logic/Segment/SegmentsExporter.php
@@ -20,6 +20,7 @@ use demosplan\DemosPlanCoreBundle\Entity\Statement\Segment;
 use demosplan\DemosPlanCoreBundle\Entity\Statement\Statement;
 use demosplan\DemosPlanCoreBundle\Logic\Export\PhpWordConfigurator;
 use demosplan\DemosPlanCoreBundle\Logic\ImageLinkConverter;
+use demosplan\DemosPlanCoreBundle\Logic\Segment\Export\ImageManager;
 use demosplan\DemosPlanCoreBundle\Services\HTMLSanitizer;
 use demosplan\DemosPlanCoreBundle\ValueObject\CellExportStyle;
 use demosplan\DemosPlanCoreBundle\ValueObject\ExportOrgaInfoHeader;
@@ -38,10 +39,6 @@ use Symfony\Contracts\Translation\TranslatorInterface;
 
 class SegmentsExporter
 {
-    private const STANDARD_DPI = 72;
-    private const STANDARD_PT_TEXT = 10;
-    private const MAX_WIDTH_INCH = 10.69;
-    private const MAX_HEIGHT_INCH = 5.42;
     /**
      * @var array<string, mixed>
      */
@@ -54,6 +51,7 @@ class SegmentsExporter
     public function __construct(
         private readonly CurrentUserInterface $currentUser,
         private readonly HTMLSanitizer $htmlSanitizer,
+        private readonly ImageManager $imageManager,
         protected readonly ImageLinkConverter $imageLinkConverter,
         Slugify $slugify,
         TranslatorInterface $translator
@@ -234,50 +232,7 @@ class SegmentsExporter
         foreach ($sortedSegments as $segment) {
             $this->addSegmentTableBody($table, $segment, $statement->getExternId());
         }
-        $this->addImages($section);
-    }
-
-    private function addImages(Section $section): void
-    {
-        // Add images after all segments of one statement.
-        $images = $this->imageLinkConverter->getImages();
-        if ([] === $images) {
-            return;
-        }
-        $imageSpaceCurrentlyUsed = 0;
-        $section->addPageBreak();
-        foreach ($images as $imageReference => $imagePath) {
-            [$width, $height] = getimagesize($imagePath);
-            [$maxWidth, $maxHeight] = $this->getMaxWidthAndHeight();
-
-            if ($width > $maxWidth) {
-                $factor = $maxWidth / $width;
-                $width = $maxWidth;
-                $height *= $factor;
-            }
-            if ($height > $maxHeight) {
-                $factor = $maxHeight / $height;
-                $height = $maxHeight;
-                $width *= $factor;
-            }
-            if ($height > $maxHeight - $imageSpaceCurrentlyUsed) {
-                $section->addPageBreak();
-            }
-            $imageSpaceCurrentlyUsed += $height + self::STANDARD_PT_TEXT * 2;
-
-            $imageStyle = [
-                'width'  => $width,
-                'height' => $height,
-                'align'  => Jc::START,
-            ];
-
-            $section->addText($imageReference);
-            $section->addBookmark($imageReference);
-            $section->addImage($imagePath, $imageStyle);
-        }
-
-        // remove already printed images
-        $this->imageLinkConverter->resetImages();
+        $this->imageManager->addImages($section);
     }
 
     protected function sortSegmentsByOrderInProcedure(array $segments): array
@@ -290,14 +245,6 @@ class SegmentsExporter
     private function compareOrderInProcedure(Segment $segmentA, Segment $segmentB): int
     {
         return $segmentA->getOrderInProcedure() - $segmentB->getOrderInProcedure();
-    }
-
-    private function getMaxWidthAndHeight(): array
-    {
-        $maxWidth = self::MAX_WIDTH_INCH * self::STANDARD_DPI;
-        $maxHeight = self::MAX_HEIGHT_INCH * self::STANDARD_DPI - self::STANDARD_PT_TEXT;
-
-        return [$maxWidth, $maxHeight];
     }
 
     private function addSegmentsTableHeader(Section $section, array $tableHeaders): Table
